backport #4039 - Fix multithreading related problems - linux makefile
git-svn-id: https://svn.osgeo.org/gdal/branches/1.8@22219 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/georaster/georaster_wrapper.cpp
@@ -1727,7 +1727,7 @@ bool GeoRasterWrapper::InitializeIO( void )
     // Allocate buffer for one raster block
     // --------------------------------------------------------------------
 
-    int nMaxBufferSize = MAX( nBlockBytes, nGDALBlockBytes );
+    long nMaxBufferSize = MAX( nBlockBytes, nGDALBlockBytes );
 
     pabyBlockBuf = (GByte*) VSIMalloc( sizeof(GByte) * nMaxBufferSize );
